pyproject.toml
[project]
name = "openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "ipykernel>=7.2.0",
    "litellm>=1.83.0",
    "ollama>=0.6.1",
    "openai-agents>=0.13.5",
    "python-dotenv>=1.2.2",
]
